Create about page backend
The about page will now be a separate HTML page, which will open when the
user clicks a link on the side-menu. This allows us to use the side-menu
for other purposes, such as to enable the user to send commands to the
server (via a form interface, rather than having to use a command style
interface).