Start websocket, send connect event
The client now starts a websocket server on port 5673, which will send a
connect event for each server instance when the `register` event is called
by each.

Using my Websocket-Listen repo, you can listen in to the raw data feed
from the websocket (provided as JSON).

(The API should probably be made a config setting in Ave, so users don't
always have to allow it through their firewalls.)