Add function to /rules and /motd buttons, new anim
Made /rules and /motd buttons actually send RULES and MOTD commands to the
server.

Updated the styles of the toggle divs (especially on hamburger menu) to
use the more animated style that spawned on the dashboard. Arguably, the
new animations are more "material" in nature (and draw the user's
attention more effectively, which is important).